New matrix strategy; EREs of desired images
Update GitHub Actions workflow to be able to better handle the upstream
tracking of desired tomcat images/tags. Matrix jobs now run and operate
on the tags described by the Extended Regular Expressions (EREs) in the
"./.github/helperScripts/desiredImages" file. In addition, this file
denotes which image should also be tracked on GitHub as the "latest"
branch.